\name{MSM-internal}
\docType{methods}
\alias{MSM-class}
\alias{MSM.linear-class}
\alias{[,MSM-method}
\alias{[,MSM.linear-method}
\alias{filter,MSM.linear-method}
\alias{smooth,MSM.linear-method}
\alias{em}
\alias{em,MSM.linear-method}

\title{Internal msmFit objects and methods}
\description{
These are not to be called by the user.\cr
If you have a specific need or for more details on msmFit, feel free to contact the author (who, even if he is very lazy, will be very happy to answer and -why not?- may start a new collaboration.)
}

\author{
	Jose A. Sanchez-Espigares, Alberto Lopez Moreno
}

\seealso{
	Overview: \code{\link{MSwM-package}} \cr
	Classes : \code{\linkS4class{MSM.lm}}, \code{\linkS4class{MSM.glm}}, \code{\linkS4class{MSM.fitted}}\cr
	Methods : \code{\link{msmFit}},\code{\link[MSwM:summary]{summary}},\code{\link[MSwM:AIC]{AIC}},\code{\link[MSwM:intervals]{intervals}},\code{\link{msmResid}}\cr
	Plot : \code{\link[MSwM:plot]{plot}},\code{\link{plotProb}},\code{\link{plotReg}},\code{\link{plotDiag}}\cr
}
\keyword{internal}
